Move extensions to CoveCore.swift
